change order_number clumn to a bigint, db save fix

+++ DeliveryDetectorServer/models.py
@@ -16,7 +16,7 @@ class BoxInfo(models.Model):
     # need a one to many field or similar, to keep track of all users using this box
 
 class OrderInfo(models.Model):
-    order_number = models.IntegerField()
+    order_number = models.BigIntegerField()
     user_name = models.CharField(max_length=200)
     seller_name = models.CharField(max_length=200)
     seller_email = models.EmailField(max_length=300)
